requirements.txt
docker>=7.0.0,<8.0.0
pexpect>=4.9.0,<5.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
click>=8.0.0,<9.0.0